+++ Modules/OTPCore/Sources/OTPCore/Clock/EpochClock.swift
@@ -3,7 +3,7 @@ import Foundation
 
 /// A clock that ticks once a second, based on the UNIX epoch.
 public protocol EpochClock {
-    func secondsPublisher() -> AnyPublisher<UInt64, Never>
+    func secondsPublisher() -> AnyPublisher<Double, Never>
     /// The current time, as most recently published.
-    var currentTime: UInt64 { get }
+    var currentTime: Double { get }
 }
